Ignore errors that prevent Jimfs.newFileSystem from working in an environment where ServiceLoader doesn't work correctly to load the SystemJimfsFileSystemProvider.
This is in relation to https://github.com/google/jimfs/issues/31... I don't entirely understand what's going on there, but it does seem like an exception being thrown from FileSystems.newFileSystem should not prevent the user from getting the FileSystem that's already been created just because it can't be cached.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=118790127
